--- src/components/Login/types.ts
@@ -1,3 +1,0 @@
-export type LoginProp = {
-  onPress: () => void;
-};
